refactor thing to noun
