+++ src/entities/interactable.ts
@@ -1,7 +1,7 @@
 import { Scene } from 'phaser';
-import Thing from '../components/thing';
+import Noun from '../components/noun';
 export default class Interactable {
-  public thing: Thing;
+  public noun: Noun;
   public sprite: Phaser.Physics.Arcade.Sprite;
   constructor(
     scene: Scene,
@@ -10,14 +10,14 @@ export default class Interactable {
     texture: string,
     staticBody = false
   ) {
-    this.thing = new Thing(texture);
+    this.noun = new Noun(texture);
     this.sprite = new Phaser.Physics.Arcade.Sprite(scene, x, y, texture);
     console.log(scene);
     scene.add.existing(this.sprite);
     scene.physics.add.existing(this.sprite, staticBody);
     this.sprite;
 
-    if (this.thing.moveable) {
+    if (this.noun.moveable) {
       this.sprite.setPushable(true);
       this.sprite.setDrag(200, 200);
     } else {
@@ -27,10 +27,10 @@ export default class Interactable {
   }
 
   update() {
-    if (this.thing.health < 2 && this.sprite?.texture.key !== 'bench-broken') {
+    if (this.noun.health < 2 && this.sprite?.texture.key !== 'bench-broken') {
       this.sprite.setTexture('bench-broken');
     }
-    if (this.thing.health < 1 && this.sprite) {
+    if (this.noun.health < 1 && this.sprite) {
       this.sprite.destroy();
     }
   }
