+++ src/systems/verbs.ts
@@ -7,7 +7,7 @@ export class Verb {
   static interact(verb: Verbs, interactable?: Interactable) {
     switch (verb) {
       case Verbs.attack:
-        interactable?.thing.takeDamage(1);
+        interactable?.noun.takeDamage(1);
         break;
     }
   }
